Refactor python code for similarity algos to use latest CAPI, get rid of python wrapper around legacy code, update CAPI and python tests
