+++ python/cugraph/CMakeLists.txt
@@ -89,7 +89,6 @@ add_subdirectory(cugraph/dask/structure)
 add_subdirectory(cugraph/internals)
 add_subdirectory(cugraph/layout)
 add_subdirectory(cugraph/linear_assignment)
-add_subdirectory(cugraph/link_prediction)
 add_subdirectory(cugraph/structure)
 add_subdirectory(cugraph/tree)
 add_subdirectory(cugraph/utilities)
