+++ python/cugraph/cugraph/link_prediction/jaccard.py
@@ -11,16 +11,45 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import cudf
-from cugraph.link_prediction import jaccard_wrapper
 from cugraph.utilities import (
     ensure_cugraph_obj_for_nx,
     df_edge_score_to_dictionary,
     renumber_vertex_pair,
 )
+import cudf
+import warnings
+
+from pylibcugraph import (
+    jaccard_coefficients as pylibcugraph_jaccard_coefficients,
+)
+from pylibcugraph import ResourceHandle
 
 
-def jaccard(input_graph, vertex_pair=None, do_expensive_check=True):
+# FIXME: Move this function to the utility module so that it can be
+# shared by other algos
+def ensure_valid_dtype(input_graph, vertex_pair):
+
+    vertex_dtype = input_graph.edgelist.edgelist_df.dtypes[0]
+    vertex_pair_dtypes = vertex_pair.dtypes
+
+    if vertex_pair_dtypes[0] != vertex_dtype or vertex_pair_dtypes[1] != vertex_dtype:
+        warning_msg = (
+            "Jaccard requires 'vertex_pair' to match the graph's 'vertex' type. "
+            f"input graph's vertex type is: {vertex_dtype} and got "
+            f"'vertex_pair' of type: {vertex_pair_dtypes}."
+        )
+        warnings.warn(warning_msg, UserWarning)
+        vertex_pair = vertex_pair.astype(vertex_dtype)
+
+    return vertex_pair
+
+
+# FIXME: 
+# 1. Be consistent with column names for output/result DataFrame
+# 2. Enforce that 'vertex_pair' is a cudf Dataframe with only columns for vertex pairs
+# 3. We need to add support for multi-column vertices
+
+def jaccard(input_graph, vertex_pair=None, do_expensive_check=False, use_weight=False):    
     """
     Compute the Jaccard similarity between each pair of vertices connected by
     an edge, or between arbitrary pairs of vertices specified by the user.
@@ -36,13 +65,11 @@ def jaccard(input_graph, vertex_pair=None, do_expensive_check=True):
     of cugraph.jaccard is different from the behavior of
     networkx.jaccard_coefficient.
 
-    This algorithm doesn't currently support datasets with vertices that
-    are not (re)numebred vertices from 0 to V-1 where V is the total number of
-    vertices as this creates isolated vertices.
-
     cugraph.jaccard, in the absence of a specified vertex pair list, will
-    use the edges of the graph to construct a vertex pair list and will
-    return the jaccard coefficient for those vertex pairs.
+    compute the two_hop_neighbors of the entire graph to construct a vertex pair
+    list and will return the jaccard coefficient for those vertex pairs. This is
+    not advisable as the vertex_pairs can grow exponentially with respect to the
+    size of the datasets
 
     networkx.jaccard_coefficient, in the absence of a specified vertex
     pair list, will return an upper triangular dense matrix, excluding
@@ -59,9 +86,9 @@ def jaccard(input_graph, vertex_pair=None, do_expensive_check=True):
     solution by doing the following:
 
     >>> from cugraph.datasets import karate
-    >>> G = karate.get_graph(download=True)
-    >>> pairs = G.get_two_hop_neighbors()
-    >>> df = cugraph.jaccard(G, pairs)
+    >>> input_graph = karate.get_graph(download=True, ignore_weights=True)
+    >>> pairs = input_graph.get_two_hop_neighbors()
+    >>> df = cugraph.jaccard(input_graph, pairs)
 
     But please remember that cugraph will fill the dataframe with the entire
     solution you request, so you'll need enough memory to store the 2-hop
@@ -72,10 +99,12 @@ def jaccard(input_graph, vertex_pair=None, do_expensive_check=True):
     ----------
     input_graph : cugraph.Graph
         cuGraph Graph instance, should contain the connectivity information
-        as an edge list (edge weights are not used for this algorithm). The
+        as an edge list (edge weights are not supported yet for this algorithm). The
         graph should be undirected where an undirected edge is represented by a
         directed edge in both direction. The adjacency list will be computed if
         not already present.
+
+        This implementation only supports undirected, unweighted Graph.
 
     vertex_pair : cudf.DataFrame, optional (default=None)
         A GPU dataframe consisting of two columns representing pairs of
@@ -84,9 +113,8 @@ def jaccard(input_graph, vertex_pair=None, do_expensive_check=True):
         current implementation computes the jaccard coefficient for all
         adjacent vertices in the graph.
 
-    do_expensive_check: bool (default=True)
-        When set to True, check if the vertices in the graph are (re)numbered
-        from 0 to V-1 where V is the total number of vertices.
+    use_weight : bool, optional (default=False)
+        Currently not supported
 
     Returns
     -------
@@ -96,10 +124,10 @@ def jaccard(input_graph, vertex_pair=None, do_expensive_check=True):
         relative to the adjacency list, or that given by the specified vertex
         pairs.
 
-        df['source'] : cudf.Series
-            The source vertex ID (will be identical to first if specified).
-        df['destination'] : cudf.Series
-            The destination vertex ID (will be identical to second if
+        df['first'] : cudf.Series
+            The first vertex ID of each pair (will be identical to first if specified).
+        df['second'] : cudf.Series
+            The second vertex ID of each pair (will be identical to second if
             specified).
         df['jaccard_coeff'] : cudf.Series
             The computed jaccard coefficient between the first and the second
@@ -108,55 +136,72 @@ def jaccard(input_graph, vertex_pair=None, do_expensive_check=True):
     Examples
     --------
     >>> from cugraph.datasets import karate
-    >>> G = karate.get_graph(download=True)
-    >>> df = cugraph.jaccard(G)
+    >>> from cugraph import jaccard
+    >>> input_graph = karate.get_graph(download=True, ignore_weights=True)
+    >>> df = jaccard(input_graph)
 
     """
-    if do_expensive_check:
-        if not input_graph.renumbered:
-            input_df = input_graph.edgelist.edgelist_df[["src", "dst"]]
-            max_vertex = input_df.max().max()
-            expected_nodes = cudf.Series(range(0, max_vertex + 1, 1)).astype(
-                input_df.dtypes[0]
-            )
-            nodes = (
-                cudf.concat([input_df["src"], input_df["dst"]])
-                .unique()
-                .sort_values()
-                .reset_index(drop=True)
-            )
-            if not expected_nodes.equals(nodes):
-                raise ValueError("Unrenumbered vertices are not supported.")
+
+    print ('use_weight =', use_weight)
+
+    warnings.warn("do_expensive_check is deprecated since it is no longer needed",
+                DeprecationWarning)
 
     if input_graph.is_directed():
         raise ValueError("Input must be an undirected Graph.")
-    if type(vertex_pair) == cudf.DataFrame:
-        vertex_pair = renumber_vertex_pair(input_graph, vertex_pair)
-    elif vertex_pair is not None:
-        raise ValueError("vertex_pair must be a cudf dataframe")
 
-    df = jaccard_wrapper.jaccard(input_graph, None, vertex_pair)
+    if vertex_pair is None:
+        # Call two_hop neighbor of the entire graph
+        vertex_pair = input_graph.get_two_hop_neighbors()
+
+    v_p_num_col = len(vertex_pair.columns)
+
+    if isinstance(vertex_pair, cudf.DataFrame):
+        vertex_pair = renumber_vertex_pair(input_graph, vertex_pair)
+        vertex_pair = ensure_valid_dtype(input_graph, vertex_pair)
+        src_col_name = vertex_pair.columns[0]
+        dst_col_name = vertex_pair.columns[1]
+        first = vertex_pair[src_col_name]
+        second = vertex_pair[dst_col_name]
+
+    elif vertex_pair is not None:
+        raise ValueError("vertex_pair must be a cudf Dataframe")
+
+    first, second, jaccard_coeff = pylibcugraph_jaccard_coefficients(
+        resource_handle=ResourceHandle(),
+        graph=input_graph._plc_graph,
+        first=first,
+        second=second,
+        use_weight=use_weight,
+        do_expensive_check=False,
+    )
 
     if input_graph.renumbered:
-        df = input_graph.unrenumber(df, "first")
-        df = input_graph.unrenumber(df, "second")
+        vertex_pair = input_graph.unrenumber(vertex_pair, src_col_name, preserve_order=True)
+        vertex_pair = input_graph.unrenumber(vertex_pair, dst_col_name, preserve_order=True)
+
+
+    if v_p_num_col == 2:
+        # single column vertex
+        vertex_pair = vertex_pair.rename(
+            columns={src_col_name: "first", dst_col_name: "second"}
+        )
+
+    df = vertex_pair
+    df["jaccard_coeff"] = cudf.Series(jaccard_coeff)
 
     return df
 
 
-def jaccard_coefficient(G, ebunch=None, do_expensive_check=True):
+def jaccard_coefficient(G, ebunch=None, do_expensive_check=False):
     """
     For NetworkX Compatability.  See `jaccard`
 
-    NOTE: This algorithm doesn't currently support datasets with vertices that
-    are not (re)numebred vertices from 0 to V-1 where V is the total number of
-    vertices as this creates isolated vertices.
-
     Parameters
     ----------
-    graph : cugraph.Graph
+    G : cugraph.Graph
         cuGraph Graph instance, should contain the connectivity information
-        as an edge list (edge weights are not used for this algorithm). The
+        as an edge list (edge weights are not supported yet for this algorithm). The
         graph should be undirected where an undirected edge is represented by a
         directed edge in both direction. The adjacency list will be computed if
         not already present.
@@ -167,6 +212,9 @@ def jaccard_coefficient(G, ebunch=None, do_expensive_check=True):
         given vertex pairs.  If the vertex_pair is not provided then the
         current implementation computes the jaccard coefficient for all
         adjacent vertices in the graph.
+
+    do_expensive_check : bool, optional (default=False)
+        Currently not supported
 
     Returns
     -------
