+++ python/cugraph/cugraph/link_prediction/overlap.py
@@ -11,28 +11,92 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from cugraph.link_prediction import overlap_wrapper
-import cudf
 from cugraph.utilities import (
     ensure_cugraph_obj_for_nx,
     df_edge_score_to_dictionary,
     renumber_vertex_pair,
 )
+import cudf
+import warnings
+
+from pylibcugraph import (
+    overlap_coefficients as pylibcugraph_overlap_coefficients,
+)
+from pylibcugraph import ResourceHandle
 
 
-def overlap_coefficient(G, ebunch=None, do_expensive_check=True):
+# FIXME: Move this function to the utility module so that it can be
+# shared by other algos
+def ensure_valid_dtype(input_graph, vertex_pair):
+
+    vertex_dtype = input_graph.edgelist.edgelist_df.dtypes[0]
+    vertex_pair_dtypes = vertex_pair.dtypes
+
+    if vertex_pair_dtypes[0] != vertex_dtype or vertex_pair_dtypes[1] != vertex_dtype:
+        warning_msg = (
+            "Overlap requires 'vertex_pair' to match the graph's 'vertex' type. "
+            f"input graph's vertex type is: {vertex_dtype} and got "
+            f"'vertex_pair' of type: {vertex_pair_dtypes}."
+        )
+        warnings.warn(warning_msg, UserWarning)
+        vertex_pair = vertex_pair.astype(vertex_dtype)
+
+    return vertex_pair
+
+
+def overlap_coefficient(G, ebunch=None, use_weight=False):
     """
     For NetworkX Compatability.  See `overlap`
 
-    NOTE: This algorithm doesn't currently support datasets with vertices that
-    are not (re)numebred vertices from 0 to V-1 where V is the total number of
-    vertices as this creates isolated vertices.
+    Parameters
+    ----------
+    G : cugraph.Graph
+        cuGraph Graph instance, should contain the connectivity information
+        as an edge list (edge weights are not supported yet for this algorithm). The
+        graph should be undirected where an undirected edge is represented by a
+        directed edge in both direction. The adjacency list will be computed if
+        not already present.
 
+    ebunch : cudf.DataFrame, optional (default=None)
+        A GPU dataframe consisting of two columns representing pairs of
+        vertices. If provided, the Overlap coefficient is computed for the
+        given vertex pairs.  If the vertex_pair is not provided then the
+        current implementation computes the overlap coefficient for all
+        adjacent vertices in the graph.
+
+    use_weight : bool, optional (default=False)
+        Currently not supported
+
+    Returns
+    -------
+    df  : cudf.DataFrame
+        GPU data frame of size E (the default) or the size of the given pairs
+        (first, second) containing the overlap weights. The ordering is
+        relative to the adjacency list, or that given by the specified vertex
+        pairs.
+
+        ddf['first']: dask_cudf.Series
+            The first vertex ID of each pair (will be identical to first if specified).
+        ddf['second']: dask_cudf.Series
+            The second vertex ID of each pair (will be identical to second if
+            specified).
+        ddf['overlap_coeff']: dask_cudf.Series
+            The computed overlap coefficient between the first and the second
+            vertex ID.
+
+    Examples
+    --------
+    >>> from cugraph.datasets import karate
+    >>> from cugraph import overlap_coefficient
+    >>> G = karate.get_graph(download=True, ignore_weights=True)
+    >>> df = overlap_coefficient(G)
     """
     vertex_pair = None
 
     G, isNx = ensure_cugraph_obj_for_nx(G)
 
+    # FIXME: What is the logic behind this since the docstrings mention that 'G' and
+    # 'ebunch'(if not None) are respectively of type cugraph.Graph and cudf.DataFrame?
     if isNx is True and ebunch is not None:
         vertex_pair = cudf.DataFrame(ebunch)
 
@@ -45,8 +109,12 @@ def overlap_coefficient(G, ebunch=None, do_expensive_check=True):
 
     return df
 
+# FIXME: 
+# 1. Be consistent with column names for output/result DataFrame
+# 2. Enforce that 'vertex_pair' is a cudf Dataframe with only columns for vertex pairs
+# 3. We need to add support for multi-column vertices
 
-def overlap(input_graph, vertex_pair=None, do_expensive_check=True):
+def overlap(G, vertex_pair=None, use_weight=False):
     """
     Compute the Overlap Coefficient between each pair of vertices connected by
     an edge, or between arbitrary pairs of vertices specified by the user.
@@ -58,25 +126,28 @@ def overlap(input_graph, vertex_pair=None, do_expensive_check=True):
     neighbors. If first is specified but second is not, or vice versa, an
     exception will be thrown.
 
-    NOTE: This algorithm doesn't currently support datasets with vertices that
-    are not (re)numebred vertices from 0 to V-1 where V is the total number of
-    vertices as this creates isolated vertices.
+    cugraph.overlap, in the absence of a specified vertex pair list, will
+    compute the two_hop_neighbors of the entire graph to construct a vertex pair
+    list and will return the overlap coefficient for those vertex pairs. This is
+    not advisable as the vertex_pairs can grow exponentially with respect to the
+    size of the datasets
 
     Parameters
     ----------
-    input_graph : cugraph.Graph
+    G : cugraph.Graph
         cuGraph Graph instance, should contain the connectivity information
-        as an edge list (edge weights are not used for this algorithm). The
+        as an edge list (edge weights are not supported yet for this algorithm). The
         adjacency list will be computed if not already present.
+
+        This implementation only supports undirected, unweighted Graph.
 
     vertex_pair : cudf.DataFrame, optional (default=None)
         A GPU dataframe consisting of two columns representing pairs of
         vertices. If provided, the overlap coefficient is computed for the
         given vertex pairs, else, it is computed for all vertex pairs.
 
-    do_expensive_check: bool (default=True)
-        When set to True, check if the vertices in the graph are (re)numbered
-        from 0 to V-1 where V is the total number of vertices.
+    use_weight : bool, optional (default=False)
+        Currently not supported
 
     Returns
     -------
@@ -98,35 +169,54 @@ def overlap(input_graph, vertex_pair=None, do_expensive_check=True):
     Examples
     --------
     >>> from cugraph.datasets import karate
-    >>> G = karate.get_graph(download=True)
-    >>> df = cugraph.overlap(G)
+    >>> from cugraph import overlap
+    >>> G = karate.get_graph(download=True, ignore_weights=True)
+    >>> df = overlap(G)
 
     """
-    if do_expensive_check:
-        if not input_graph.renumbered:
-            input_df = input_graph.edgelist.edgelist_df[["src", "dst"]]
-            max_vertex = input_df.max().max()
-            expected_nodes = cudf.Series(range(0, max_vertex + 1, 1)).astype(
-                input_df.dtypes[0]
-            )
-            nodes = (
-                cudf.concat([input_df["src"], input_df["dst"]])
-                .unique()
-                .sort_values()
-                .reset_index(drop=True)
-            )
-            if not expected_nodes.equals(nodes):
-                raise ValueError("Unrenumbered vertices are not supported.")
 
-    if type(vertex_pair) == cudf.DataFrame:
-        vertex_pair = renumber_vertex_pair(input_graph, vertex_pair)
+    if G.is_directed():
+        raise ValueError("Input must be an undirected Graph.")
+
+
+    if vertex_pair is None:
+        # Call two_hop neighbor of the entire graph
+        vertex_pair = G.get_two_hop_neighbors()
+
+    v_p_num_col = len(vertex_pair.columns)
+
+    if isinstance(vertex_pair, cudf.DataFrame):
+        vertex_pair = renumber_vertex_pair(G, vertex_pair)
+        vertex_pair = ensure_valid_dtype(G, vertex_pair)
+        src_col_name = vertex_pair.columns[0]
+        dst_col_name = vertex_pair.columns[1]
+        first = vertex_pair[src_col_name]
+        second = vertex_pair[dst_col_name]
+
     elif vertex_pair is not None:
         raise ValueError("vertex_pair must be a cudf dataframe")
 
-    df = overlap_wrapper.overlap(input_graph, None, vertex_pair)
+    
+    first, second, overlap_coeff = pylibcugraph_overlap_coefficients(
+        resource_handle=ResourceHandle(),
+        graph=G._plc_graph,
+        first=first,
+        second=second,
+        use_weight=use_weight,
+        do_expensive_check=False,
+    )
 
-    if input_graph.renumbered:
-        df = input_graph.unrenumber(df, "first")
-        df = input_graph.unrenumber(df, "second")
+    if G.renumbered:
+        vertex_pair = G.unrenumber(vertex_pair, src_col_name, preserve_order=True)
+        vertex_pair = G.unrenumber(vertex_pair, dst_col_name, preserve_order=True)
+
+    if v_p_num_col == 2:
+        # single column vertex
+        vertex_pair = vertex_pair.rename(
+            columns={src_col_name: "first", dst_col_name: "second"}
+        )
+
+    df = vertex_pair
+    df["overlap_coeff"] = cudf.Series(overlap_coeff)
 
     return df
