+++ python/cugraph/cugraph/tests/link_prediction/test_jaccard.py
@@ -11,6 +11,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+#FIXME: Can we use global variables for column names instead of hardcoded ones?
+
 import gc
 
 import pytest
@@ -20,9 +22,9 @@ import cudf
 import cugraph
 from cugraph.datasets import netscience
 from cugraph.testing import utils, UNDIRECTED_DATASETS
-from cugraph.experimental import jaccard as exp_jaccard
+from cugraph import jaccard
 from cudf.testing import assert_series_equal, assert_frame_equal
-from cugraph.experimental import jaccard_coefficient as exp_jaccard_coefficient
+from cugraph import jaccard_coefficient
 
 
 print("Networkx version : {} ".format(nx.__version__))
@@ -38,7 +40,9 @@ def setup_function():
 # =============================================================================
 # Helper functions
 # =============================================================================
-def compare_jaccard_two_hop(G, Gnx, edgevals=True):
+
+
+def compare_jaccard_two_hop(G, Gnx, use_weight=False):
     """
     Compute both cugraph and nx jaccard after extracting the two hop neighbors
     from G and compare both results
@@ -49,29 +53,28 @@ def compare_jaccard_two_hop(G, Gnx, edgevals=True):
         .reset_index(drop=True)
     )
 
-    nx_pairs = list(pairs.to_records(index=False))
-    preds = nx.jaccard_coefficient(Gnx, nx_pairs)
-    nx_coeff = []
-    for u, v, p in preds:
-        # print(u, " ", v, " ", p)
-        nx_coeff.append(p)
     df = cugraph.jaccard(G, pairs)
     df = df.sort_values(by=["first", "second"]).reset_index(drop=True)
-    if not edgevals:
-        # experimental jaccard currently only supports unweighted graphs
-        df_exp = exp_jaccard(G, pairs)
-        df_exp = df_exp.sort_values(by=["first", "second"]).reset_index(drop=True)
-        assert_frame_equal(df, df_exp, check_dtype=False, check_like=True)
 
-    assert len(nx_coeff) == len(df)
-    for i in range(len(df)):
-        diff = abs(nx_coeff[i] - df["jaccard_coeff"].iloc[i])
-        assert diff < 1.0e-6
+    if not use_weight:
+        nx_pairs = list(pairs.to_records(index=False))
+        preds = nx.jaccard_coefficient(Gnx, nx_pairs)
+        nx_coeff = []
+        for u, v, p in preds:
+            nx_coeff.append(p)
+
+        assert len(nx_coeff) == len(df)
+        for i in range(len(df)):
+            diff = abs(nx_coeff[i] - df["jaccard_coeff"].iloc[i])
+            assert diff < 1.0e-6
+    else:
+        #FIXME: compare results against resultset api
+        pass
 
 
-def cugraph_call(benchmark_callable, graph_file, edgevals=False, input_df=None):
+def cugraph_call(benchmark_callable, graph_file, input_df=None, use_weight=False):
     G = cugraph.Graph()
-    G = graph_file.get_graph(ignore_weights=not edgevals)
+    G = graph_file.get_graph(ignore_weights=not use_weight)
 
     # If no vertex_pair is passed as input, 'cugraph.jaccard' will
     # compute the 'jaccard_similarity' with the two_hop_neighbor of the
@@ -82,7 +85,7 @@ def cugraph_call(benchmark_callable, graph_file, edgevals=False, input_df=None):
     vertex_pair = vertex_pair[["first", "second"]]
 
     # cugraph Jaccard Call
-    df = benchmark_callable(cugraph.jaccard, G, vertex_pair=vertex_pair)
+    df = benchmark_callable(cugraph.jaccard, G, vertex_pair=vertex_pair, use_weight=use_weight)
 
     df = df.sort_values(["first", "second"]).reset_index(drop=True)
 
@@ -144,27 +147,34 @@ def read_csv(request):
 
 
 @pytest.mark.sg
-def test_jaccard(read_csv, gpubenchmark):
+@pytest.mark.parametrize("use_weight", [False, True])
+def test_jaccard(read_csv, gpubenchmark, use_weight):
 
     M_cu, M, graph_file = read_csv
-    cu_src, cu_dst, cu_coeff = cugraph_call(gpubenchmark, graph_file, input_df=M_cu)
-    nx_src, nx_dst, nx_coeff = networkx_call(M)
+    cu_src, cu_dst, cu_coeff = cugraph_call(gpubenchmark, graph_file, input_df=M_cu, use_weight=use_weight)
+    if not use_weight:
+        nx_src, nx_dst, nx_coeff = networkx_call(M)
 
-    # Calculating mismatch
-    err = 0
-    tol = 1.0e-06
+        # Calculating mismatch
+        err = 0
+        tol = 1.0e-06
 
-    assert len(cu_coeff) == len(nx_coeff)
-    for i in range(len(cu_coeff)):
-        if abs(cu_coeff[i] - nx_coeff[i]) > tol * 1.1:
-            err += 1
+        assert len(cu_coeff) == len(nx_coeff)
+        for i in range(len(cu_coeff)):
+            if abs(cu_coeff[i] - nx_coeff[i]) > tol * 1.1:
+                err += 1
 
-    print("Mismatches:  %d" % err)
-    assert err == 0
+        print("Mismatches:  %d" % err)
+        assert err == 0
+    else:
+        #FIXME: compare weighted jaccard results against resultset api
+        pass
+    
 
 
 @pytest.mark.sg
-def test_directed_graph_check(read_csv):
+@pytest.mark.parametrize("use_weight", [False, True])
+def test_directed_graph_check(read_csv, use_weight):
     _, M, _ = read_csv
 
     cu_M = cudf.DataFrame()
@@ -180,7 +190,7 @@ def test_directed_graph_check(read_csv):
     vertex_pair = cu_M[["src_0", "src_1", "dst_0", "dst_1"]]
     vertex_pair = vertex_pair[:5]
     with pytest.raises(ValueError):
-        cugraph.jaccard(G1, vertex_pair)
+        cugraph.jaccard(G1, vertex_pair, use_weight)
 
 
 @pytest.mark.sg
@@ -192,52 +202,43 @@ def test_nx_jaccard_time(read_csv, gpubenchmark):
 
 @pytest.mark.sg
 @pytest.mark.parametrize("graph_file", [netscience])
-@pytest.mark.skip(reason="Skipping because this datasets is unrenumbered")
-def test_jaccard_edgevals(gpubenchmark, graph_file):
+@pytest.mark.parametrize("use_weight", [False, True])
+def test_jaccard_edgevals(gpubenchmark, graph_file, use_weight):
     dataset_path = netscience.get_path()
     M = utils.read_csv_for_nx(dataset_path)
     M_cu = utils.read_csv_file(dataset_path)
     cu_src, cu_dst, cu_coeff = cugraph_call(
-        gpubenchmark, netscience, edgevals=True, input_df=M_cu
+        gpubenchmark, netscience, input_df=M_cu, use_weight=use_weight
     )
-    nx_src, nx_dst, nx_coeff = networkx_call(M)
+    if not use_weight:
+        nx_src, nx_dst, nx_coeff = networkx_call(M)
 
-    # Calculating mismatch
-    err = 0
-    tol = 1.0e-06
+        # Calculating mismatch
+        err = 0
+        tol = 1.0e-06
 
-    assert len(cu_coeff) == len(nx_coeff)
-    for i in range(len(cu_coeff)):
-        if abs(cu_coeff[i] - nx_coeff[i]) > tol * 1.1:
-            err += 1
+        assert len(cu_coeff) == len(nx_coeff)
+        for i in range(len(cu_coeff)):
+            if abs(cu_coeff[i] - nx_coeff[i]) > tol * 1.1:
+                err += 1
 
-    print("Mismatches:  %d" % err)
-    assert err == 0
+        print("Mismatches:  %d" % err)
+        assert err == 0
+    else:
+        #FIXME: compare results against resultset api
+        pass
 
 
 @pytest.mark.sg
-def test_jaccard_two_hop(read_csv):
+@pytest.mark.parametrize("use_weight", [False, True])
+def test_jaccard_two_hop(read_csv, use_weight):
 
     _, M, graph_file = read_csv
 
     Gnx = nx.from_pandas_edgelist(M, source="0", target="1", create_using=nx.Graph())
-    G = graph_file.get_graph(ignore_weights=True)
+    G = graph_file.get_graph(ignore_weights= not use_weight)
 
-    compare_jaccard_two_hop(G, Gnx)
-
-
-@pytest.mark.sg
-def test_jaccard_two_hop_edge_vals(read_csv):
-
-    _, M, graph_file = read_csv
-
-    Gnx = nx.from_pandas_edgelist(
-        M, source="0", target="1", edge_attr="weight", create_using=nx.Graph()
-    )
-
-    G = graph_file.get_graph()
-
-    compare_jaccard_two_hop(G, Gnx, edgevals=True)
+    compare_jaccard_two_hop(G, Gnx, use_weight)
 
 
 @pytest.mark.sg
@@ -252,10 +253,8 @@ def test_jaccard_nx(read_csv):
     ebunch = M_cu.rename(columns={"0": "first", "1": "second"})
     ebunch = ebunch[["first", "second"]]
     cg_j = cugraph.jaccard_coefficient(Gnx, ebunch=ebunch)
-    cg_j_exp = exp_jaccard_coefficient(Gnx, ebunch=ebunch)
 
     assert len(nv_js) > len(cg_j)
-    assert len(nv_js) > len(cg_j_exp)
 
     # FIXME:  Nx does a full all-pair Jaccard.
     # cuGraph does a limited 1-hop Jaccard
@@ -280,51 +279,18 @@ def test_jaccard_multi_column(read_csv):
     vertex_pair = cu_M[["src_0", "src_1", "dst_0", "dst_1"]]
     vertex_pair = vertex_pair[:5]
 
-    df_res = cugraph.jaccard(G1, vertex_pair)
-    df_plc_exp = exp_jaccard(G1, vertex_pair)
+    df_multi_col_res = cugraph.jaccard(G1, vertex_pair)
+    jaccard_multi_col_res = df_multi_col_res["jaccard_coeff"].sort_values().reset_index(drop=True)
 
-    df_plc_exp = df_plc_exp.rename(
-        columns={
-            "0_src": "0_source",
-            "0_dst": "0_destination",
-            "1_src": "1_source",
-            "1_dst": "1_destination",
-        }
-    )
+    print(df_multi_col_res)
 
-    jaccard_res = df_res["jaccard_coeff"].sort_values().reset_index(drop=True)
-    jaccard_plc_exp = df_plc_exp["jaccard_coeff"].sort_values().reset_index(drop=True)
-    assert_series_equal(jaccard_res, jaccard_plc_exp)
 
     G2 = cugraph.Graph()
     G2.from_cudf_edgelist(cu_M, source="src_0", destination="dst_0")
-    df_exp = cugraph.jaccard(G2, vertex_pair[["src_0", "dst_0"]])
+    df_single_col_res = cugraph.jaccard(G2, vertex_pair[["src_0", "dst_0"]])
 
     # Calculating mismatch
-    actual = df_res.sort_values("0_first").reset_index()
-    expected = df_exp.sort_values("first").reset_index()
+    actual = df_multi_col_res.sort_values("0_src").reset_index()
+    expected = df_single_col_res.sort_values("first").reset_index()
+
     assert_series_equal(actual["jaccard_coeff"], expected["jaccard_coeff"])
-
-
-@pytest.mark.sg
-def test_weighted_exp_jaccard():
-    karate = UNDIRECTED_DATASETS[0]
-    G = karate.get_graph()
-    with pytest.raises(ValueError):
-        exp_jaccard(G)
-
-    G = karate.get_graph(ignore_weights=True)
-    use_weight = True
-    with pytest.raises(ValueError):
-        exp_jaccard(G, use_weight=use_weight)
-
-
-@pytest.mark.sg
-def test_invalid_datasets_jaccard():
-    karate = UNDIRECTED_DATASETS[0]
-    df = karate.get_edgelist()
-    df = df.add(1)
-    G = cugraph.Graph(directed=False)
-    G.from_cudf_edgelist(df, source="src", destination="dst")
-    with pytest.raises(ValueError):
-        cugraph.jaccard(G)
