+++ python/pylibcugraph/pylibcugraph/__init__.py
@@ -87,6 +87,13 @@ from pylibcugraph.generate_rmat_edgelist import generate_rmat_edgelist
 
 from pylibcugraph.generate_rmat_edgelists import generate_rmat_edgelists
 
+from pylibcugraph.jaccard_coefficients import jaccard_coefficients
+
+from pylibcugraph.overlap_coefficients import overlap_coefficients
+
+from pylibcugraph.sorensen_coefficients import sorensen_coefficients
+
+
 from pylibcugraph import exceptions
 
 __version__ = "23.10.00"
